+++ services/transcode/playout_mux_test.go
@@ -9,7 +9,7 @@ import (
 )
 
 func Test_GenerateFFmpegParamsForPlayoutMux(t *testing.T) {
-	const golden = `-progress pipe:1 -hide_banner -i root/BERG_TS01_ISRAEL_VOD.mxf -i root/BERG_TS01_ISRAEL_VOD-nor.wav -i root/BERG_TS01_ISRAEL_VOD-eng.wav -i root/BERG_TS01_ISRAEL_VOD-fin.wav -filter_complex [1:a]aresample=48000,channelsplit=channel_layout=stereo[nor_l][nor_r];[nor_l]asplit=10[nor_l_copy_0][nor_l_copy_1][nor_l_copy_2][nor_l_copy_3][nor_l_copy_4][nor_l_copy_5][nor_l_copy_6][nor_l_copy_7][nor_l_copy_8][nor_l_copy_9];[nor_r]asplit=3[nor_r_copy_0][nor_r_copy_1][nor_r_copy_2];[2:a]aresample=48000,channelsplit=channel_layout=stereo[eng_l][eng_r];[3:a]aresample=48000[fin_l] -map 0:v -map [nor_l_copy_0] -map [nor_r_copy_0] -map [nor_l_copy_1] -map [nor_r_copy_1] -map [nor_l_copy_2] -map [nor_r_copy_2] -map [eng_l] -map [eng_r] -map [nor_l_copy_3] -map [nor_l_copy_4] -map [fin_l] -map [nor_l_copy_5] -map [nor_l_copy_6] -map [nor_l_copy_7] -map [nor_l_copy_8] -map [nor_l_copy_9] -c:v copy -c:a pcm_s24le -y something/something.mxf`
+	const golden = `-progress pipe:1 -hide_banner -i root/BERG_TS01_ISRAEL_VOD.mxf -i root/BERG_TS01_ISRAEL_VOD-nor.wav -i root/BERG_TS01_ISRAEL_VOD-eng.wav -i root/BERG_TS01_ISRAEL_VOD-fin.wav -filter_complex [1:a]aresample=48000,channelsplit=channel_layout=stereo[nor_l][nor_r];[nor_l]asplit=10[nor_l_copy_0][nor_l_copy_1][nor_l_copy_2][nor_l_copy_3][nor_l_copy_4][nor_l_copy_5][nor_l_copy_6][nor_l_copy_7][nor_l_copy_8][nor_l_copy_9];[nor_r]asplit=3[nor_r_copy_0][nor_r_copy_1][nor_r_copy_2];[2:a]aresample=48000,channelsplit=channel_layout=stereo[eng_l][eng_r];[3:a]aresample=48000,pan=1c|c0=c0[fin_l] -map 0:v -map [nor_l_copy_0] -map [nor_r_copy_0] -map [nor_l_copy_1] -map [nor_r_copy_1] -map [nor_l_copy_2] -map [nor_r_copy_2] -map [eng_l] -map [eng_r] -map [nor_l_copy_3] -map [nor_l_copy_4] -map [fin_l] -map [nor_l_copy_5] -map [nor_l_copy_6] -map [nor_l_copy_7] -map [nor_l_copy_8] -map [nor_l_copy_9] -c:v copy -c:a pcm_s24le -y something/something.mxf`
 
 	const root = "root/"
 	const outputPath = "something/something.mxf"
